Change default value for 'width' setting to NULL
Select2 should not set a specific width by default, as this breaks the responsiveness of the plugin. Setting the default value to NULL will make Select2.prototype._resolveWidth() return NULL as well, which in turn prevents Select2.prototype._placeContainer() from applying a CSS-width to the container which would make it unresponsive.