Updated front page features
This adds correct links to the linked text on the front page under
the features section. There is still one panel which needs to be
written, the last one that includes the AJAX text again, before it
can really be considered finished.

The wording around the Bootstrap theme was adjusted pending further
discussion about how that is going to be handled.