Adjusted the home page text
This adds the final box, which I've chosen to be for the browser
support (though that can be changed).

The last two boxes on the first row have also been swapped, partially
because the text sizes now match the second row, but also because
the AJAX support should be more prominent than the theme support.