Lock Firefox builds to Linux
Right now Sauce Labs is not triggering builds for Firefox, and that
is causing the automated testing to fail. This might be because
there is no "latest" version logged for Firefox on Sauce Labs, but
it could also be becuase we aren't giving it a platform to look for.

This was also done for Opera as right now Sauce Labs is running
tests on Linux anyway.