Properly disable `input` handlers in IE
Previously we were only disabling the `input` handler when it was
triggered, which caused a race condition within the `keyup` handlers
which also was triggered by the `input` event. This fixes the issue by
also unbinding the `input` handlers within the `keyup` handler, to avoid
running into the race condition.

Thanks to @Eckankar for pointing out the race condition that still
existed in
https://github.com/select2/select2/commit/66ae2ad1d5a5b6c983b17a342550c61d9cd15a82

This closes https://github.com/select2/select2/issues/3300