Added pull request and issue templates
The issue template is largely based on the questions included in the
canned comment that I use for bugs, which is mostly covered in the
contributing guide. Even though there are a lot of lines, most of it is
just ticking checkboxes.

The issue template is focused on bug reports, as those are considerably
more common than feature requests.

The pull request template was designed to be pretty slim, we haven't had
many issues with these in the past.