Better detection of aborted ajax queries
On Chrome 50, `$request.status` is a number, not a string thus the aborted query detection failed. For the end-user this resulted in seeing the 'The results could not be loaded.' message when he typed too fast for too long, instead of the correct 'Searching…' message.

This commit check `$request.status` both as a number and as a string, keeping backward compatibility while fixing the case for Chrome.