Fix cleanup of scroll positioning handler
If the Select2 DOM elements are removed from the DOM or moved to another parent element while the select2 dropdown is open, the scroll event would not get detached from the correct elements. This breaks scrolling and causes a memory leak. This patch fixes the problem by storing the elements to which event handlers were bound.