Add a fetchAll option
fetchAll means that your ajax data source is called only once with a
null search term. The server is assumed to return all possible options.
The results are cached and then searched using the default matcher as if
they had been local results.

This saves a lot of calls into the server and means you don't need to
implement search on the server, unless you want to search much larger
data sets, in which case you'd probably return nothing for a null search
term (rather than everything), and not use fetchAll.

This basically implements this feature request:

    https://github.com/select2/select2/issues/5400

But I didn't think there was much point ever searching the server with a
non-null search term if we're caching (since normally all options are
retrieved as soon as someone even clicks on the widget), so I made it a
"fetchAll" option rather than a "cache" option.